graphql-file structure cleaned up
