+++ service/src/graphql/resolvers/index.ts
@@ -1,0 +1,5 @@
+import { todoMutation } from "./mutation/todo.mutations";
+import { todoQuery } from "./query/todo.query";
+
+
+export const resolvers = [todoQuery, todoMutation]
