+++ service/src/graphql/schemas/index.ts
@@ -1,0 +1,3 @@
+import { todoTypeDefs } from "./todo.schema";
+
+export const typeDefs = [todoTypeDefs]
